% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gene_set.R
\name{add_gene_set}
\alias{add_gene_set}
\title{add_gene_set}
\usage{
add_gene_set(geneset, name = NA, desc = NA, genelist = NA)
}
\arguments{
\item{geneset}{is a GeneSet object for pathway annotation}

\item{name}{is the name of the gene set to be added}

\item{desc}{is a short description of the gene set}

\item{genelist}{is a list of the gene ids to be added}
}
\description{
add_gene_set function Add a gene set to a GeneSet object.
}
\details{
This function adds a single new gene set to an existing GeneSet object. It returns
             the same (now modified) GeneSet object.
}
\examples{
dontrun{
  mygeneset = make_gene_set()
  mygeneset = add_gene_set(mygeneset, name="MyPathway", 
                           desc="A pathway describing an important thing",
                           genelist=c("geneA","geneB","geneX","geneY"))

}

}
