% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_redundant_rows.R
\name{average_redundant_rows}
\alias{average_redundant_rows}
\title{average_redundant_rows}
\usage{
average_redundant_rows(data)
}
\arguments{
\item{data}{is a...}
}
\description{
average_redundant_rows function description is...
}
\examples{
dontrun{


}

}
