% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottleneck_overlap.R
\name{bottleneck_overlap}
\alias{bottleneck_overlap}
\title{bottleneck_overlap}
\usage{
bottleneck_overlap(set_1, set_2, percentage = 0.2)
}
\arguments{
\item{set_1}{is...}

\item{set_2}{is...}

\item{percentage}{default is .2}
}
\description{
bottleneck_overlap function description is...
}
\examples{
dontrun{


}

}
