% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_predictor.R
\name{build_predictor}
\alias{build_predictor}
\title{build_predictor}
\usage{
build_predictor(fit, data, level = NULL, signature = NULL,
  genesymbols_from = TRUE, genesymbols_to = NULL, browser = F)
}
\arguments{
\item{fit}{is...}

\item{data}{is...}

\item{level}{default is NULL}

\item{signature}{default is NULL}

\item{genesymbols_from}{default is TRUE}

\item{genesymbols_to}{default is NULL}

\item{browser}{default is FALSE}
}
\description{
build_predictor function description is...
}
\examples{
dontrun{


}

}
