% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_data_table_1.R
\name{clinical_data_table_1}
\alias{clinical_data_table_1}
\title{clinical_data_table_1}
\usage{
clinical_data_table_1(patientlist, clinical_data)
}
\arguments{
\item{patientlist}{is a list...}

\item{clinical_data}{is...}
}
\description{
clinical_data_table_1 function description is...
}
\examples{
dontrun{


}

}
