% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_gene_sets.R
\name{combine_gene_sets}
\alias{combine_gene_sets}
\title{combine_gene_sets}
\usage{
combine_gene_sets(genesetA, genesetB, tagA = NULL, tagB = NULL)
}
\arguments{
\item{genesetA}{is...}

\item{genesetB}{is...}

\item{tagA}{default is NULL}

\item{tagB}{default is NULL}
}
\description{
combine_gene_sets function description is...
}
\examples{
dontrun{


}

}
