% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compare_column_names}
\alias{compare_column_names}
\title{compare_column_names}
\usage{
compare_column_names(abundance_data1, abundance_data2)
}
\arguments{
\item{abundance_data1}{first data object of class 'abundance_data'}

\item{abundance_data2}{second data object of class 'abundance_data'}
}
\value{
reports number of columns in common
}
\description{
This function compares column names between two abundance data objects
}
