% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{compare_id_names}
\alias{compare_id_names}
\title{compare_id_names}
\usage{
compare_id_names(abundance_data1, abundance_data2)
}
\arguments{
\item{abundance_data1}{first data object of class 'abundance_data'}

\item{abundance_data2}{second data object of class 'abundance_data'}
}
\description{
This function compares identifier names between two abundance data objects
}
