% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_coefficients.R
\name{consensus_coefficients}
\alias{consensus_coefficients}
\title{consensus_coefficients}
\usage{
consensus_coefficients(fitcv, level, return.all = T,
  filter_count = NULL, meanweight = NULL, subsample = NULL)
}
\arguments{
\item{fitcv}{is...}

\item{level}{is...}

\item{return.all}{default is TRUE}

\item{filter_count}{default is NULL}

\item{meanweight}{default is NULL}

\item{subsample}{default is NULL}
}
\description{
consensus_coefficients function description is...
}
\examples{
dontrun{


}

}
