% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_model.R
\name{consensus_model}
\alias{consensus_model}
\title{consensus_model}
\usage{
consensus_model(model_list, return.all = T, filter_count = NULL)
}
\arguments{
\item{model_list}{is...}

\item{return.all}{default is TRUE}

\item{filter_count}{default is NULL}
}
\description{
consensus_model function description is...
}
\examples{
dontrun{


}

}
