% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_comparison_enrichment.R
\name{correlation_comparison_enrichment}
\alias{correlation_comparison_enrichment}
\title{correlation_comparison_enrichment}
\usage{
correlation_comparison_enrichment(
  geneset,
  abundance,
  set1,
  set2,
  mapping_column = NA,
  tag = NA,
  mode = "original"
)
}
\description{
# internal function to calculate enrichment in differences in correlation between two groups
# access through the leapr wrapper
}
