% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_expression.R
\name{coxph_expression}
\alias{coxph_expression}
\title{coxph_expression}
\usage{
coxph_expression(survivalmatrix, statusvar, timevar, startcolumn)
}
\arguments{
\item{survivalmatrix}{is...}

\item{statusvar}{is...}

\item{timevar}{is...}

\item{startcolumn}{is...}
}
\description{
coxph_expression function description is...
}
\examples{
dontrun{


}

}
