% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_random_subset.R
\name{coxph_random_subset}
\alias{coxph_random_subset}
\title{coxph_random_subset}
\usage{
coxph_random_subset(survivalmatrix, statusvar, timevar, startcolumn,
  size = 5, ntimes = 50)
}
\arguments{
\item{survivalmatrix}{is...}

\item{statusvar}{is...}

\item{timevar}{is...}

\item{startcolumn}{is...}

\item{size}{default is 5}

\item{ntimes}{default is 50}
}
\description{
coxph_random_subset function description...
}
\examples{
dontrun{


}

}
