% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph_subset.R
\name{coxph_subset}
\alias{coxph_subset}
\title{coxph_subset}
\usage{
coxph_subset(survivalmatrix, statusvar = "status", timevar = "time",
  columns, rows, verbose = F)
}
\arguments{
\item{survivalmatrix}{is...}

\item{statusvar}{default is 'status'}

\item{timevar}{defult is 'time'}

\item{columns}{is...}

\item{rows}{is...}

\item{verbose}{default is FALSE}
}
\description{
coxph_subset function description is...
}
\examples{
dontrun{


}

}
