% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_matrix.R
\name{cross_validation_matrix}
\alias{cross_validation_matrix}
\title{cross_validation_matrix}
\usage{
cross_validation_matrix(validation_dataset, survivaldata, timevar,
  statusvar, fitcv, data, consensus_foldx = 25,
  consensus_maxsteps = 100, meanweight = NULL, ...)
}
\arguments{
\item{validation_dataset}{is...}

\item{survivaldata}{is...}

\item{timevar}{is...}

\item{statusvar}{is...}

\item{fitcv}{is...}

\item{data}{is...}

\item{consensus_foldx}{default is 25}

\item{consensus_maxsteps}{default is 100}

\item{meanweight}{default is NULL}

\item{...}{is...}
}
\description{
cross_validation_matrix function description is...
}
\examples{
dontrun{


}

}
