% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_with_holdout.R
\name{cross_validation_with_holdout}
\alias{cross_validation_with_holdout}
\title{cross_validation_with_holdout}
\usage{
cross_validation_with_holdout(data, survivaldata, timevar, statusvar,
  consensus_holdout = 0.5, validation_holdout = 0.3333,
  validation_dataset = NA, foldx = 5, consensus_foldx = 25,
  consensus_maxsteps = 100, consensus_level = 40,
  consensus_count_filter = 4, load.saved.file = T, ...)
}
\arguments{
\item{data}{is...}

\item{survivaldata}{is...}

\item{timevar}{is...}

\item{statusvar}{is...}

\item{consensus_holdout}{default is 0.5}

\item{validation_holdout}{default is 0.3333}

\item{validation_dataset}{default is NA}

\item{foldx}{default is 5}

\item{consensus_foldx}{default is 25}

\item{consensus_maxsteps}{default is 100}

\item{consensus_level}{defaul is 40}

\item{consensus_count_filter}{default is 4}

\item{load.saved.file}{Logical indicating whether or not to load saved file. Default is TRUE}

\item{...}{is...}
}
\description{
cross_validation_with_holdout function description is...
}
\examples{
dontrun{


}

}
