% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_significance_summary.R
\name{cv_significance_summary}
\alias{cv_significance_summary}
\title{cv_significance_summary}
\usage{
cv_significance_summary(mintable, results_struct, pthresh = 0.05,
  return.percentage = T)
}
\arguments{
\item{mintable}{is...}

\item{results_struct}{is...}

\item{pthresh}{defaults to 0.05}

\item{return.percentage}{is a logical indicating whether percentage should be returned. Defaults to TRUE}
}
\description{
cv_significance_summary function description is...
}
\examples{
dontrun{


}

}
