% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_enrichment_in_relationships.R
\name{difference_enrichment_in_relationships}
\alias{difference_enrichment_in_relationships}
\title{difference_enrichment_in_relationships}
\usage{
difference_enrichment_in_relationships(
  geneset,
  relationships1,
  relationships2,
  idmap = NA,
  tag = NA,
  mode = "original"
)
}
\arguments{
\item{geneset}{is...}

\item{relationships1}{is...}

\item{relationships2}{is...}

\item{idmap}{defaults to an NA}

\item{tag}{defaults to an NA}
}
\description{
difference_enrichment_in_relationships function description is...
}
\examples{
dontrun{


}

}
