% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_comparison.R
\name{enrichment_comparison}
\alias{enrichment_comparison}
\title{enrichment_comparison}
\usage{
enrichment_comparison(
  geneset,
  abundance,
  mapping_column = NULL,
  abundance_column = NULL,
  fdr = 0,
  matchset = NULL,
  longform = F,
  sample_comparison,
  min_p_threshold = NULL,
  tag = NA,
  sample_n = NULL
)
}
\description{
# access through leapr wrapper
}
