% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_abundance.R
\name{enrichment_in_abundance}
\alias{enrichment_in_abundance}
\title{enrichment_in_abundance}
\usage{
enrichment_in_abundance(
  geneset,
  abundance,
  mapping_column = NULL,
  abundance_column = NULL,
  fdr = 0,
  matchset = NULL,
  longform = F,
  sample_comparison = NULL,
  min_p_threshold = NULL,
  tag = NA,
  sample_n = NULL
)
}
\arguments{
\item{geneset}{is a GeneSet object for pathway annotation}

\item{mapping_column}{Is an optional character string, a column name of \code{abundance}, for specifying gene mapping (e.g. for phosphoproteomics data). Defaults to NULL}

\item{abundance_column}{Is a character vector composed of column names from \code{abundance}, that identifies which conditions should be tested.}

\item{fdr}{A numerical value which specifies how many times to randomly sample genes, defaults to 0.}

\item{matchset}{defaults to NULL}

\item{sample_comparison}{Is a character vector composed of column names from \code{abundance}, that identifies which conditions should be compared. Defaults to NULL}

\item{min_p_threshold}{Is a numeric value, a lower p-value threshold for returning results, defaults to NULL.}

\item{sample_n}{If set specifies the number of randomizations to perform to calculate an FDR [warning: time intensive], default is NA}

\item{Is}{a \emph{mxn} matrix of abundance data (protein, gene, etc.), with \emph{m} gene names (rows) and \emph{n} sample/condition (columns).}

\item{a}{prefix tag (optional) for specifying source of data default is NA}
}
\description{
Enrichment in abundance calculates enrichment in pathways by the difference in abundance of the pathway members.
}
\details{
Case 1: Single condition compatison (sample_comparison = NULL). This usage applies a T test to

assess the significance of the abundance of pathway members relative to

genes/proteins for a condition or conditions that are not in the pathway for a single condition

Case 2: Case/control comparison (sample_comparison is specified. Compare the abundance of

the members in the pathway for one set of conditions (abundance_column - which can be a vector) versus

another set of conditions (sample_comparison - also a vector).

These are calculated for every pathway in geneset.

The significance of this is based on the hypothesis that pathway activity is related to abdundance,

so that if the members of a praticular pathway are higher abundance than everything else, then it

can be considered more 'active' (Case 1). Or in Case 2, if the abundance in pathway members is higher

in one set of conditions (case) relative to in another set of conditions (controls) then it is more

'active' in that case.

Note that the use of the T test to statistically test the difference between the two sets of abundance

measures is very simple and may be subject to caveats (e.g. low number of comparisons, not-normally

distributed abundance values, etc.)

Providing a number for \code{sample_n} specifies the number of times the pathway membership will be

randomized to calculate a fasle discovery rate for the enrichment. This is a time-intensive option

since it requires running the enrichment sample_n+1 times.

The interpretation of results from Case 1 to Case 2 should vary considerably.
}
\examples{
dontrun{
        library(LEAP)

        # read in the example abundance data
        data("protdata")

        # read in the pathways
        data("ncipid")

        # read in the patient groups
        data("short_list")
        data("longlist")

        #in this example we lump a bunch of patients together (the 'short survivors') and compare them to another group (the 'long survivors')
        protdata.enrichment.svl = enrichment_in_abundance(ncipid, protdata, abundance_column=shortlist, sample_comparison=longlist)

        #another application is to compare just one patient against another (this would be the equivalent of comparing one time point to another)
        protdata.enrichment.svl.ovo = enrichment_in_abundance(ncipid, protdata, abundance_column=shortlist[1], sample_comparison=longlist[1])

}



}
