% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_order.R
\name{enrichment_in_order}
\alias{enrichment_in_order}
\title{enrichment_in_order}
\usage{
enrichment_in_order(
  geneset,
  background = NULL,
  minsize = 5,
  mapping_column = NULL,
  abundance_column = NULL,
  randomize = F
)
}
\arguments{
\item{background}{Is a \emph{mxn} matrix of gene expression data, with \emph{m} gene names (rows) and \emph{n} sample/condition (columns).}

\item{minsize}{Is the minimum size of a gene set that will be considered, defaults to 5}

\item{mapping_column}{Is an optional character string, a column name of \code{abundance}, for specifying gene mapping (e.g. for phosphoproteomics data). Defaults to NULL}

\item{abundance_column}{Is a character vector composed of column names from \code{background}, that ...??.}

\item{randomize}{is a logical, defaults to FALSE}

\item{genesets}{is a GeneSet object for pathway annotation}
}
\description{
Enrichment in order is a wrapper function for enrichment_in_groups where 'method' argument is set to "ks" and "targets" is set to NA
}
