% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_pathway.R
\name{enrichment_in_pathway}
\alias{enrichment_in_pathway}
\title{enrichment_in_pathway}
\usage{
enrichment_in_pathway(
  geneset,
  abundance,
  mapping_column = NULL,
  abundance_column = NULL,
  fdr = 0,
  matchset = NULL,
  longform = F,
  min_p_threshold = NULL,
  tag = NA,
  sample_n = NULL
)
}
\description{
Enrichment in pathway is a wrapper function for enrichment_in_abundance where 'sample_comparison' argument is set to NULL
}
