% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_relationships.R
\name{enrichment_in_relationships}
\alias{enrichment_in_relationships}
\title{enrichment_in_relationships}
\usage{
enrichment_in_relationships(
  geneset,
  relationships,
  idmap = NA,
  mode = "original",
  silence_try_errors = TRUE
)
}
\arguments{
\item{geneset}{List of pathways in gmt format}

\item{relationships}{table of relationship information, e.g. correlation}

\item{idmap}{list of identifiers to use for mapping, the names of the items should agree with names of features in matrix}

\item{mode}{allowable values are 'original' or anything else}

\item{silence_try_errors}{boolean to silence errors}
}
\value{
table of enrichment statistics
}
\description{
enrichment_in_relationships function description is a general way to determine if a pathway
is enriched in relationships (interactions, correlation) between its members
# access through leapr wrapper
}
