% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_sets.R
\name{enrichment_in_sets}
\alias{enrichment_in_sets}
\title{enrichment_in_sets}
\usage{
enrichment_in_sets(
  geneset,
  targets = NULL,
  background = NULL,
  minsize = 5,
  mapping_column = NULL,
  randomize = F
)
}
\description{
Enrichment in sets is a wrapper function for enrichment_in_groups where 'method' argument is set to "fishers" and "abundance_column" is set to NA
# access through leapr wrapper
}
