% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_redundancy_matrix.R
\name{enrichment_redundancy_matrix}
\alias{enrichment_redundancy_matrix}
\title{enrichment_redundancy_matrix}
\usage{
enrichment_redundancy_matrix(geneset, dataframe = NA,
  enrichment_results = NA, significance_threshold = NA,
  pathway_list = NA, method = "jaccard")
}
\arguments{
\item{geneset}{is...}

\item{dataframe}{defaults to NA}

\item{enrichment_results}{defaults to NA}

\item{significance_threshold}{defaults to NA}

\item{pathway_list}{defaults to NA}

\item{method}{default is 'jaccard'}
}
\description{
enrichment_redundancy_matrix function description is...
}
\examples{
dontrun{


}

}
