% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_wrapper_JEM.R
\name{enrichment_wrapper}
\alias{enrichment_wrapper}
\title{enrichment_wrapper}
\usage{
enrichment_wrapper(geneset, enrichment_method, ...)
}
\arguments{
\item{geneset}{is...}

\item{enrichment_method}{is...}

\item{datamatrix}{is...}

\item{id_column}{is...}
}
\description{
enrichment_wrapper is a wrapper function that consolidates multiple enrichment methods.
}
\examples{
dontrun{


}

}
