% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gene_set.R
\name{extract_gene_set}
\alias{extract_gene_set}
\title{extract_gene_set}
\usage{
extract_gene_set(geneset, setlist)
}
\arguments{
\item{geneset}{is...}

\item{setlist}{is...}
}
\description{
extract_gene_set function description is...
}
\examples{
dontrun{


}

}
