% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_coxph.R
\name{gene_set_coxph}
\alias{gene_set_coxph}
\title{gene_set_coxph}
\usage{
gene_set_coxph(data, survivaldata, timevar, statusvar, geneset,
  min.geneset.size = 6, num.repeats = 10, randomize.order = F,
  randomize.genes = F, match.set.size = F, extra.features = NULL,
  sigthresh = 0.05, translation_table = NULL,
  training.set.ratio = 0.75, sum.sig = F, sig.vote = F)
}
\arguments{
\item{data}{is...}

\item{survivaldata}{is...}

\item{timevar}{is...}

\item{statusvar}{is...}

\item{geneset}{is...}

\item{min.geneset.size}{default is 6}

\item{num.repeats}{default is 10}

\item{randomize.order}{logical indicates whether order should be randomized, default is FALSE}

\item{randomize.genes}{logical indicates whether genes should be randomized, default is FALSE}

\item{match.set.size}{logical, defaults to FALSE}

\item{extra.features}{defaults to NULL}

\item{sigthresh}{defaults to 0.05}

\item{translation_table}{defaults to NULL}

\item{training.set.ratio}{defaults to 0.75}

\item{sum.sig}{logical, defaults to FALSE}

\item{sig.vote}{logical, defaults to FALSE}
}
\description{
gene_set_coxph function description is...
}
\examples{
dontrun{


}

}
