% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_heatmap.R
\name{gene_set_heatmap}
\alias{gene_set_heatmap}
\title{gene_set_heatmap}
\usage{
gene_set_heatmap(inmatrix, colsets = NA, filter = 1.5)
}
\arguments{
\item{inmatrix}{is...}

\item{colsets}{default is an NA}

\item{filter}{defualt is 1.5}
}
\description{
gene_set_heatmap function description is...
}
\examples{
dontrun{


}

}
