% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_overlap.R
\name{gene_set_overlap}
\alias{gene_set_overlap}
\title{gene_set_overlap}
\usage{
gene_set_overlap(inmatrix, colsets = NA, filter = 1.5)
}
\arguments{
\item{inmatrix}{is...}

\item{colsets}{defaults to an NA}

\item{filter}{defaults to 1.5}
}
\description{
gene_set_overlap function description is...
}
\examples{
dontrun{


}

}
