% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set_rpf.R
\name{gene_set_rpf}
\alias{gene_set_rpf}
\title{gene_set_rpf}
\usage{
gene_set_rpf(data, survivaldata, timevar, statusvar, geneset,
  load.saved.file = T, only.load.files = F, gsprefix = "",
  min.geneset.size = 6, max.steps = 30, num.repeats = 0,
  pre.pvalue.threshold = NULL, validation.data = NULL,
  make.validata = T, collect.validation.pvalues = F,
  p.value.threshold = 0.05, randomize.order = F, randomize.genes = F,
  match.set.size = F, training.set.ratio = 0.5,
  extra.features = NULL, generator.mode = F, save.cv.pvalues = F,
  translation_table = NULL, signature.level = 20, sig.vote = F, ...)
}
\arguments{
\item{data}{is...}

\item{survivaldata}{is...}

\item{timevar}{is...}

\item{statusvar}{is...}

\item{geneset}{is...}

\item{load.saved.file}{logical, defaults to TRUE}

\item{only.load.files}{logical, defaults to FALSE}

\item{gsprefix}{defaults to ''}

\item{min.geneset.size}{defaults to 6}

\item{max.steps}{defaults to 30}

\item{num.repeats}{defaults to 0}

\item{pre.pvalue.threshold}{defaults to NULL}

\item{validation.data}{defaults to NULL}

\item{make.validata}{defaults to TRUE}

\item{collect.validation.pvalues}{defaults to FALSE}

\item{p.value.threshold}{defaults to 0.05}

\item{randomize.order}{logical, defaults to FALSE}

\item{randomize.genes}{logical, defaults to FALSE}

\item{match.set.size}{logical, defaults to FALSE}

\item{training.set.ratio}{defaults to 0.5}

\item{extra.features}{defaults to NULL}

\item{generator.mode}{logical, defaults to FALSE}

\item{save.cv.pvalues}{logical defaults to FALSE}

\item{translation_table}{defaults to NULL}

\item{signature.level}{defaults to 20}

\item{sig.vote}{logical defaults to FALSE}

\item{...}{is ...}
}
\description{
gene_set_rpf function description is...
}
\examples{
dontrun{


}

}
