% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_pca.R
\name{geneset_pca}
\alias{geneset_pca}
\title{geneset_pca}
\usage{
geneset_pca(geneset, datatable, minsize = 5, ncomp = 1)
}
\arguments{
\item{geneset}{is...}

\item{datatable}{is...}

\item{minsize}{defaults to 5}

\item{ncomp}{defaults to 1}
}
\description{
geneset_pca function description is...
}
\examples{
dontrun{


}

}
