% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_membership_matrix.R
\name{group_membership_matrix}
\alias{group_membership_matrix}
\title{group_membership_matrix}
\usage{
group_membership_matrix(geneset, pathwayname, sources, mode = "original")
}
\arguments{
\item{geneset}{is...}

\item{pathwayname}{is...}

\item{sources}{is...}

\item{mode}{defaults to 'original'}
}
\description{
group_membership_matrix function description is...
}
\examples{
dontrun{


}

}
