% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kinasesubstrates}
\alias{kinasesubstrates}
\title{Kinase substrate lists}
\format{
A list with 4 items
\describe{
\item{names}{The names of the kinases}
\item{desc}{Short description of the kinase}
\item{sizes}{Length of the substrate list}
\item{matrix}{Substrate list for the kinase}
}
}
\source{
PhosphositePlus
}
\usage{
data(kinasesubstrates)
}
\description{
Kinase substrate lists
}
\keyword{datasets}
