% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_enrichment_vector.R
\name{ks_enrichment_vector}
\alias{ks_enrichment_vector}
\title{ks_enrichment_vector}
\usage{
ks_enrichment_vector(groupgeneids, backgroundgenes, map_column = 0,
  sort_col = 1)
}
\arguments{
\item{groupgeneids}{is...}

\item{backgroundgenes}{is...}

\item{map_column}{defaults to 0}

\item{sort_col}{defaults to 1}
}
\description{
ks_enrichment_vector function description is...
}
\examples{
dontrun{


}

}
