% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{longlist}
\alias{longlist}
\title{Long list of patient samples}
\format{
An object of class \code{character} of length 37.
}
\usage{
data(longlist)
}
\description{
Long list of patient samples
}
\keyword{datasets}
