% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gene_set.R
\name{make_gene_set}
\alias{make_gene_set}
\title{make_gene_set}
\usage{
make_gene_set(name, desc, genelist)
}
\arguments{
\item{name}{is...}

\item{desc}{is...}

\item{genelist}{is...}
}
\description{
make_gene_set function description is...
}
\examples{
dontrun{


}

}
