% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_gene_set_matrix.R
\name{make_gene_set_matrix}
\alias{make_gene_set_matrix}
\title{make_gene_set_matrix}
\usage{
make_gene_set_matrix(database)
}
\arguments{
\item{database}{is...}
}
\description{
make_gene_set_matrix function description is...
}
\examples{
dontrun{


}

}
