% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mo_krbpaths}
\alias{mo_krbpaths}
\title{Multi-Omic KEGG, Reactome, BioCarta Pathways}
\format{
A list with 4 items
\describe{
\item{names}{The names of the pathways}
\item{desc}{Short description of the pathways}
\item{sizes}{Number of genes in the signaling pathways}
\item{matrix}{Matrix containing the genes in the pathways with multi-omic prefixes}
}
}
\source{
MSIGDB
}
\usage{
mo_krbpaths
}
\description{
Multi-Omic KEGG, Reactome, BioCarta Pathways
}
\keyword{datasets}
