% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb}
\alias{msigdb}
\title{MSigDB Gene Lists}
\format{
A list with 4 items
\describe{
\item{names} The names of the signaling pathways
\item{desc}
\item{sizes}
\item{matrix}
}
}
\source{
MSIGDB
}
\usage{
msigdb
}
\description{
MSigDB Gene Lists
}
\keyword{datasets}
