% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncipid}
\alias{ncipid}
\title{NCI Gene lists}
\format{
A list with 4 items
\describe{
\item{names}{The names of the signaling pathways}
\item{desc}{Short description of the pathways}
\item{sizes}{Number of genes in the signaling pathways}
\item{matrix}{Matrix containing the genes in the pathways}
}
}
\source{
NCIPID
}
\usage{
data(ncipid)
}
\description{
A list of pathways and the genes that comprise these pathways
}
\keyword{datasets}
