% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordered_list_enrichment_from_matrix.R
\name{ordered_list_enrichment_from_matrix}
\alias{ordered_list_enrichment_from_matrix}
\title{ordered_list_enrichment_from_matrix}
\usage{
ordered_list_enrichment_from_matrix(geneset, datamatrix,
  starting_col = NULL, ending_col = NULL, idcolumn = NULL,
  threshold = 0.02, background_list = NULL,
  enrichment_threshold = 0.05)
}
\arguments{
\item{geneset}{is...}

\item{datamatrix}{is...}

\item{starting_col}{defaults to NULL}

\item{ending_col}{defaults to NULL}

\item{idcolumn}{defaults to NULL}

\item{threshold}{defaults to 0.02}

\item{background_list}{defaults to NULL}

\item{enrichment_threshold}{defaults to 0.05}
}
\description{
ordered_list_enrichment_from_matrix function description is...
}
\examples{
dontrun{


}

}
