% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_overlap_enrichment.R
\name{pairwise_overlap_enrichment}
\alias{pairwise_overlap_enrichment}
\title{pairwise_overlap_enrichment}
\usage{
pairwise_overlap_enrichment(geneset, datamatrix, condition1, condition2,
  mapping_column = NULL, subsample_components = NULL)
}
\arguments{
\item{geneset}{is...}

\item{datamatrix}{is...}

\item{condition1}{is...}

\item{condition2}{is...}

\item{mapping_column}{defaults to NULL}

\item{subsample_components}{defaults to NULL}
}
\description{
pairwise_overlap_enrichment function description is...
}
\examples{
dontrun{


}

}
