% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_sets_enrichment.R
\name{partition_sets_enrichment}
\alias{partition_sets_enrichment}
\title{partition_sets_enrichment}
\usage{
partition_sets_enrichment(partition_sets, geneset, abundance,
  mapping_column = NULL, abundance_column = NULL,
  sample_comparison = NULL, background_comparison = NULL)
}
\arguments{
\item{partition_sets}{is...}

\item{geneset}{is...}

\item{abundance}{is...}

\item{mapping_column}{defaults to NULL}

\item{abundance_column}{defaults to NULL}

\item{sample_comparison}{defaults to NULL}

\item{background_comparison}{defaults to NULL}
}
\description{
partiton_sets_enrichment function description is...
}
\examples{
dontrun{


}

}
