% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_sets_ksenrich.R
\name{partition_sets_ksenrich}
\alias{partition_sets_ksenrich}
\title{partition_sets_ksenrich}
\usage{
partition_sets_ksenrich(partition_sets, geneset, abundance,
  mapping_column = NULL)
}
\arguments{
\item{partition_sets}{is...}

\item{geneset}{is...}

\item{abundance}{is...}

\item{mapping_column}{defaults to NULL}
}
\description{
partition_sets_ksenrich function description is...
}
\examples{
dontrun{


}

}
