% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_model_predictor_matrix.R
\name{pathway_model_predictor_matrix}
\alias{pathway_model_predictor_matrix}
\title{pathway_model_predictor_matrix}
\usage{
pathway_model_predictor_matrix(pathfit, validation_data = NULL)
}
\arguments{
\item{pathfit}{is...}

\item{validation_data}{defaults to NULL}
}
\description{
pathway_model_predictor_matrix function description is...
}
\examples{
dontrun{


}

}
