% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSetPCA.R
\name{pca_cluster_distinctness}
\alias{pca_cluster_distinctness}
\title{pca cluster distinctness}
\usage{
pca_cluster_distinctness(geneset, datatable, groups, labels = NULL,
  includeall = F, mode = "original")
}
\arguments{
\item{geneset}{is...description of function argument}

\item{datatable}{is...description of function argument}

\item{groups}{is...description of function argument}

\item{labels}{is...description of function argument}

\item{includeall}{is...description of function argument}

\item{mode}{is...description of function argument}
}
\description{
pca cluster distinctness
}
