% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_enrichment_in_groups.R
\name{permute_enrichment_in_groups}
\alias{permute_enrichment_in_groups}
\title{permute_enrichment_in_groups}
\usage{
permute_enrichment_in_groups(
  geneset,
  targets = NULL,
  background = NULL,
  method = "fishers",
  minsize = 5,
  mapping_column = NULL,
  abundance_column = NULL,
  ntimes = 100
)
}
\arguments{
\item{geneset}{is a list of four vectors, gene names, gene descriptions, gene sizes and a matrix...??}

\item{targets}{defaults to NULL}

\item{background}{Is a \emph{mxn} matrix of gene expression data, with \emph{m} gene names (rows) and \emph{n} sample/condition (columns).}

\item{method}{defaults to 'fishers'}

\item{minsize}{defaults to 5}

\item{mapping_column}{Is a character string, a column name of \code{background}, that...??.}

\item{abundance_column}{defaults to NULL}

\item{ntimes}{defaults to 100}
}
\description{
permute_enrichment_in_groups function description is...
}
