% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_expression.R
\name{plot_expression}
\alias{plot_expression}
\title{plot_expression}
\usage{
plot_expression(ratios, rowname, pfactor = 1, xvector = NA,
  xfactor = 0, title = NA, condrange = c(), trendline = FALSE,
  trendcolor = "red", color = "grey", legend = FALSE, yr = NA,
  mode = "new")
}
\arguments{
\item{ratios}{is...}

\item{rowname}{is...}

\item{pfactor}{defaults to 1}

\item{xvector}{defaults to NA}

\item{xfactor}{defaults to 0}

\item{title}{defaults to NA}

\item{condrange}{is...}

\item{trendline}{defaults to FALSE}

\item{trendcolor}{defaults to 'red'}

\item{color}{defaults to 'grey'}

\item{legend}{defaults to FALSE}

\item{yr}{defaults to NA}

\item{mode}{defaults to "new"}
}
\description{
plot_expression function description is...
}
\examples{
dontrun{


}

}
