% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_kaplan_meier.R
\name{plot_model_kaplan_meier}
\alias{plot_model_kaplan_meier}
\title{plot_model_kaplan_meier}
\usage{
plot_model_kaplan_meier(fit, data, level, signature = NULL,
  prediction = NULL, training.set = NULL, group.ratios = c(0.5, 0.5),
  plotthis = T, extra = NULL, return.survival = F,
  genesymbols_from = T, ...)
}
\arguments{
\item{fit}{is...}

\item{data}{is...}

\item{level}{is...}

\item{signature}{defaults to NULL}

\item{prediction}{defaults to NULL}

\item{training.set}{defaults to NULL}

\item{group.ratios}{defaults to c(0.5,0.5)}

\item{plotthis}{defaults to TRUE}

\item{extra}{defaults to NULL}

\item{return.survival}{defaults to FALSE}

\item{genesymbols_from}{defaults to TRUE}

\item{...}{is...}
}
\description{
plto_model_kaplan_meier function description is...
}
\examples{
dontrun{


}

}
