% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_model.R
\name{print.model}
\alias{print.model}
\title{print.model}
\usage{
\method{print}{model}(fit, level = NULL, pvalues = NULL)
}
\arguments{
\item{fit}{is...}

\item{level}{defaults to NULL}

\item{pvalues}{defaults to NULL}
}
\description{
print.model function description is...
}
\examples{
dontrun{


}

}
