% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progressive_enrichment.R
\name{progressive_enrichment}
\alias{progressive_enrichment}
\title{progressive_enrichment}
\usage{
progressive_enrichment(geneset, rankedlist, decreasing = T,
  ntimes = 10, nstep = 10, startat = 10, endat = 0.5)
}
\arguments{
\item{geneset}{is...}

\item{rankedlist}{is...}

\item{decreasing}{defaults to TRUE}

\item{ntimes}{defaults to 10}

\item{nstep}{defaults to 10}

\item{startat}{defaults at 10}

\item{endat}{defaults to .5}
}
\description{
progressive_enrichment function description is...
}
\examples{
dontrun{


}

}
