% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protdata}
\alias{protdata}
\title{Protein expression data}
\format{
A data frame with 1999 rows and 174 columns
}
\source{
H. Zhang et al., Integrated Proteogenomic Characterization of Human High-Grade Serous Ovarian Cancer. Cell 166, 755-765 (2016)
}
\usage{
protdata
}
\description{
A dataset containing protein expression data containing 1999 proteins and 174 patient samples. 
Values represent the logratio of the protein expression compared to a control
}
\keyword{datasets}
