% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_gene_set_overlap_db.R
\name{random_gene_set_overlap_db}
\alias{random_gene_set_overlap_db}
\title{random_gene_set_overlap_db}
\usage{
random_gene_set_overlap_db(inmatrix_1, inmatrix_2, n1, n2, rep = 100)
}
\arguments{
\item{inmatrix_1}{is...}

\item{inmatrix_2}{is...}

\item{n1}{is...}

\item{n2}{is...}

\item{rep}{defaults to 100}
}
\description{
random_gene_set_overlap_db function description is...
}
\examples{
dontrun{


}

}
