% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_gene_set_overlap_x3.R
\name{random_gene_set_overlap_x3}
\alias{random_gene_set_overlap_x3}
\title{random_gene_set_overlap_x3}
\usage{
random_gene_set_overlap_x3(list_1, list_2, list_3, n1, n2, n3, rep = 100)
}
\arguments{
\item{list_1}{is...}

\item{list_2}{is...}

\item{list_3}{is...}

\item{n1}{is...}

\item{n2}{is...}

\item{n3}{is...}

\item{rep}{defaults to 1000}
}
\description{
random_gene_set_overlap_x3 function description is...
}
\examples{
dontrun{


}

}
