% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_overlap.R
\name{random_overlap}
\alias{random_overlap}
\title{random_overlap}
\usage{
random_overlap(items, counts = c(10, 50, 100), N = 3, rep = 100)
}
\arguments{
\item{items}{is...}

\item{counts}{is...}

\item{N}{defaults to 3}

\item{rep}{defaults to 100}
}
\description{
random_overlap function description is...
}
\examples{
dontrun{


}

}
