% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_abundance_data.R
\name{read_abundance_data}
\alias{read_abundance_data}
\title{read_abundance_data}
\usage{
read_abundance_data(filename, data_type, infocol_name = NULL)
}
\arguments{
\item{filename}{the directory path to a file containing the abundance data}

\item{data_type}{a character string, either 'protdata' or 'phosphoprotdata' which describes the type of abundance data}

\item{infocol_name}{a character string describing the name of the column extra identifier info in abundance data of type 'phosphoprotdata'}
}
\description{
This function reads in abundance data
}
\examples{
dontrun{
        read_abundance_data("/OvarianSitePhosphorylation.txt", data_type = "phospho_protdata", infocol_name = "hgnc_symbol")

}

}
