% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_sets.R
\name{read_gene_sets}
\alias{read_gene_sets}
\title{read_gene_sets}
\usage{
read_gene_sets(
  gsfile,
  gene.labels = NA,
  gs.size.threshold.min = 5,
  gs.size.threshold.max = 15000
)
}
\arguments{
\item{gsfile}{is a gene set file, for example a .gmt file (gene matrix transposed file format)}

\item{gene.labels}{defaults to NA}

\item{gs.size.threshold.min}{defaults to 5}

\item{gs.size.threshold.max}{defaults to 15000}
}
\value{
gene set object

geneset list
}
\description{
read_gene_sets is a function to import external pathway database files in .gmt format
}
\examples{
gfile = system.file('extdata','h.all.v2024.1.Hs.symbols.gmt',package='leapR')
glist = read_gene_sets(gfile)

}
