% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpf_coxph_formatdata.R
\name{rpf_coxph_formatdata}
\alias{rpf_coxph_formatdata}
\title{rpf_coxph_formatdata}
\usage{
rpf_coxph_formatdata(data, survdata, timevar, statusvar,
  extra.features = NULL)
}
\arguments{
\item{data}{is...}

\item{survdata}{is...}

\item{timevar}{is...}

\item{statusvar}{is...}

\item{extra.features}{defaults to NULL}
}
\description{
rpf_coxph_formatdata function description is...
}
\examples{
dontrun{


}

}
