% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_formodel.R
\name{survival_formodel}
\alias{survival_formodel}
\title{survival_formodel}
\usage{
survival_formodel(genename, leapstruct, dataset, consensusmat)
}
\arguments{
\item{genename}{is...}

\item{leapstruct}{is...}

\item{dataset}{is...}

\item{consensusmat}{is...}
}
\description{
survival_formodel function description is...
}
\examples{
dontrun{


}

}
