% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transdata}
\alias{transdata}
\title{Transcriptomic data}
\format{
A data frame with 18632 rows and 174 columns

A data frame with 20732 rows and 70 columns - first column is protein id
}
\source{
H. Zhang et al., Integrated Proteogenomic Characterization of Human High-Grade Serous Ovarian Cancer. Cell 166, 755-765 (2016)

H. Zhang et al., Integrated Proteogenomic Characterization of Human High-Grade Serous Ovarian Cancer. Cell 166, 755-765 (2016)
}
\usage{
transdata

transdata
}
\description{
A dataset containing gene transcriptomic data containing 18632 genes and 174 patient samples. 
Values represent normalized (Z score) gene expression relative to the mean expression for that gene

A dataset containing phosphoproteomic data containing 20732 phosphosites and 69 patient samples. 
Values represent normalized (Z score) phosphorylation adjusted to the protein abundance
}
\keyword{datasets}
