% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gene_set.R
\name{add_gene_set}
\alias{add_gene_set}
\title{add_gene_set}
\usage{
add_gene_set(geneset, name = NA, desc = NA, genelist = NA)
}
\arguments{
\item{geneset}{is...}

\item{name}{default is an NA}

\item{genelist}{default is an NA}
}
\description{
add_gene_set function description...
}
\examples{
dontrun{


}

}
