% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_enrichment.R
\name{cluster_enrichment}
\alias{cluster_enrichment}
\title{cluster_enrichment}
\usage{
cluster_enrichment(database, clusters, background = NA,
  sigfilter = 0.05)
}
\arguments{
\item{database}{is...}

\item{clusters}{is...}

\item{background}{default is an NA}

\item{sigfilter}{default is .05}
}
\description{
Cluster enrichment function description...
}
\examples{
dontrun{


}

}
