% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterEnrichment.R
\name{cluster_enrichment}
\alias{cluster_enrichment}
\title{Cluster Enrichment}
\usage{
cluster_enrichment(database, clusters, background = NA,
  sigfilter = 0.05)
}
\arguments{
\item{hclust}{is...description of function argument}

\item{annotation}{is...description of function argument}

\item{cutrange}{is...description of function argument}
}
\description{
Cluster Enrichment
}
