% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_enrichment.R
\name{cluster_enrichment}
\alias{cluster_enrichment}
\title{cluster_enrichment}
\usage{
cluster_enrichment(geneset, clusters, background = NA, sigfilter = 0.05)
}
\arguments{
\item{geneset}{is a GeneSet object for pathway annotation}

\item{clusters}{is a list of clusters (gene lists) to calculate enrichment on, generally the result of the `cuttree` function}

\item{background}{is a list of genes to serve as the background for enrichment}

\item{sigfilter}{minimum significance threshold default is .05}
}
\description{
Cluster enrichment Run enrichment (Fisher's exact) on clusters (lists of identifier groups)
}
\details{
This function will calculate enrichment (Fisher's exact test for membership overlap) on

a series of lists of genes, such as from a set of clusters. The results are returned as

a list of results matrices in the order of the input clusters.
}
\examples{
\dontrun{
        library(leapr)

        # read in the example transcriptomic data
        datadir='https://github.com/pnnl/leapR/raw/refs/heads/bioc-submission/csv/'
        transdata<-readr::read_csv(paste0(datadir,'transdata.csv.gz'))|>
           tibble::column_to_rownames('...1')|>
             as.matrix()

        # read in the pathways
        data("ncipid")
        
        # for the example we will limit the number of transcripts considered - arbitrarily in this case
        transdata = transdata[1:3000,]

        # perform heirarchical clustering on the  data
        transdata.hc = hclust(dist(transdata), method="ward.D2")
        
        transdata.hc.clusters = cutree(transdata.hc, k=5)
        
        #calculates enrichment for each of the clusters individually and returns a list
        #   of enrichment results
        transdata.hc.enrichment = cluster_enrichment(geneset=ncipid, clusters=transdata.hc.clusters, background=rownames(transdata))
        
}

}
