% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_omics.R
\name{combine_omics}
\alias{combine_omics}
\title{combine_omics
Combine two or more omics matrices into one multi-omics matrix with
'tagged' ids.}
\usage{
combine_omics(omics_list, id_list = rep(NA, length(omics_list)))
}
\arguments{
\item{omics_list}{Is a list of \code{SummarizedExperiment} each with
one assay}

\item{id_list}{List of identifiers to use, in the same order as the
omics_list elements. If an element
is `NA`, then rownames are used.}
}
\value{
\code{SummarizedExperiment} with an additional assay called
 `combined`
}
\description{
combine_omics
Combine two or more omics matrices into one multi-omics matrix with
'tagged' ids.
}
\details{
This combines matrices of different omics types together and
adds prefix tags to the ids.
}
\examples{
        library(leapR)
        library(BiocFileCache)
        path <- tools::R_user_dir("leapR", which = "cache")
        bfc <- BiocFileCache(path, ask = FALSE)
        
        url <- "https://api.figshare.com/v2/file/download/56536217"
        pc <- bfcadd(bfc, "pdat", fpath = url)
        load(pc)
        
        url <- "https://api.figshare.com/v2/file/download/56536214"
        tc <- bfcadd(bfc, "tdat", fpath = url)
        load(tc)
        
        url <- "https://api.figshare.com/v2/file/download/56536211"
        phc <- bfcadd(bfc, "phdat", fpath = url)
        load(phc)

        # merge the three datasets by rows and add prefix tags for
        # different omics types
        multi_omics <- combine_omics(list(pset, tset, phset),
                    list(NA,NA,'hgnc_id'))


}
