% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_omics.R
\name{combine_omics}
\alias{combine_omics}
\title{combine_omics}
\usage{
combine_omics(omics_list, id_list = rep(NA, length(omics_list)))
}
\arguments{
\item{transcriptomics}{is a matrix of gene expression values  where rownames are ids and columns are conditions}

\item{proteomics}{is a matrix of transcript values where rownames are ids and columns are conditions}

\item{methylation}{is a matrix of methylation values where rownames are ids and columns are conditions}

\item{cnv}{is a matrix of copy number variant(CNV) values where rownames are ids and columns are conditions}

\item{phospho}{is a dataframe of phosphorylation data}

\item{proteomics_tag}{is a text prefix to be added to protein ids}

\item{transcriptomics_tag}{is a text prefix to be added to transcript ids}

\item{methylation_tag}{is a text prefix to be added to methylation ids}

\item{cnv_tag}{is a text prefix to be added to cnv ids}

\item{phospho_tag}{is a text prefix to be added to phospho ids}

\item{id_column}{is an optional column number for identifiers for the phospho data}
}
\value{
combined omics table
}
\description{
Combine two or more omics matrices into one multi-omics matrix with 'tagged' ids.
}
\details{
This combines matrices of different omics types together and adds prefix tags to the ids.
}
\examples{
        library(leapR)


        pdata <- download.file('https://figshare.com/ndownloader/files/55781147',method='libcurl',destfile='protData.rda')
        load('protData.rda')
        p <- file.remove("protData.rda")
        
        tdata <- download.file("https://figshare.com/ndownloader/files/55781153",method='libcurl',destfile='transData.rda')
        load('transData.rda')
        p <- file.remove("transData.rda")
        
        phdata<-download.file('https://figshare.com/ndownloader/files/55781120',method='libcurl',destfile = 'phosData.rda')
        #phosphodata<-read.csv("phdata",check.names=FALSE,row.names=1)
        load('phosData.rda')
        p <- file.remove('phosData.rda')# read in the example protein data
        

        # merge the two datasets by rows and add prefix tags for different omics types
        multi_omics = combine_omics(list(pset, tset, phset))


}
