% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_omics.R
\name{combine_omics}
\alias{combine_omics}
\title{combine_omics}
\usage{
combine_omics(
  proteomics = NA,
  transcriptomics = NA,
  methylation = NA,
  cnv = NA,
  phospho = NA,
  proteomics_tag = "prot_",
  transcriptomics_tag = "txn_",
  methylation_tag = "meth_",
  cnv_tag = "cnv_",
  phospho_tag = "phospho_",
  id_column = NA
)
}
\arguments{
\item{proteomics}{is a matrix of transcript values where rownames are ids and columns are conditions}

\item{transcriptomics}{is a matrix of gene expression values  where rownames are ids and columns are conditions}

\item{methylation}{is a matrix of methylation values where rownames are ids and columns are conditions}

\item{cnv}{is a matrix of copy number variant(CNV) values where rownames are ids and columns are conditions}

\item{phospho}{is a dataframe of phosphorylation data}

\item{proteomics_tag}{is a text prefix to be added to protein ids}

\item{transcriptomics_tag}{is a text prefix to be added to transcript ids}

\item{methylation_tag}{is a text prefix to be added to methylation ids}

\item{cnv_tag}{is a text prefix to be added to cnv ids}

\item{phospho_tag}{is a text prefix to be added to phospho ids}

\item{id_column}{is an optional column number for identifiers for the phospho data}
}
\description{
Combine two or more omics matrices into one multi-omics matrix with 'tagged' ids.
}
\details{
This combines matrices of different omics types together and adds prefix tags to the ids.
}
\examples{
dontrun{
        library(leapr)

        # read in the example protein data
        data("protdata")
        # read in the example transcriptomics data
        data("transdata")
        
        # merge the two datasets by rows and add prefix tags for different omics types
        multi_omics = combine_omics(proteomics=protdata, transcriptomics=transdata)

}

}
