% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_omics.R
\name{combine_omics}
\alias{combine_omics}
\title{combine_omics}
\usage{
combine_omics(omics_list, id_list = rep(NA, length(omics_list)))
}
\arguments{
\item{omics_list}{List of \code{ExpressionSet} objects that contain the data to join with data
type in the `Annotation` slot}

\item{id_list}{List of identifers to use, in the same order as the omics_list elements. If an element
is `NA`, then rownames are used.}
}
\value{
combined omics \code{ExpressionSet}
}
\description{
Combine two or more omics matrices into one multi-omics matrix with 'tagged' ids.
}
\details{
This combines matrices of different omics types together and adds prefix tags to the ids.
}
\examples{
        library(leapR)


        pdata <- download.file('https://api.figshare.com/v2/file/download/55781147',method='libcurl',destfile='protData.rda')
        load('protData.rda')
        p <- file.remove("protData.rda")
        
        tdata <- download.file("https://api.figshare.com/v2/file/download/55781153",method='libcurl',destfile='transData.rda')
        load('transData.rda')
        p <- file.remove("transData.rda")
        
        phdata<-download.file('https://api.figshare.com/v2/file/download/55781120',method='libcurl',destfile = 'phosData.rda')
        #phosphodata<-read.csv("phdata",check.names=FALSE,row.names=1)
        load('phosData.rda')
        p <- file.remove('phosData.rda')# read in the example protein data
        

        # merge the three datasets by rows and add prefix tags for different omics types
        multi_omics = combine_omics(list(pset, tset, phset), list(NA,NA,'hgnc_id'))


}
