% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_comparison_enrichment.R
\name{correlation_comparison_enrichment}
\alias{correlation_comparison_enrichment}
\title{correlation_comparison_enrichment}
\usage{
correlation_comparison_enrichment(
  geneset,
  abundance,
  set1,
  set2,
  mapping_column = NA,
  tag = NA,
  mode = "original"
)
}
\arguments{
\item{geneset}{pathway to use for enrichment}

\item{abundance}{abundance matrix}

\item{set1}{first set to use}

\item{set2}{second set to use}

\item{mapping_column}{Column to use for id mapping}

\item{tag}{Tag to add}

\item{mode}{to use, default is 'original'}
}
\value{
data frame with enrichment results
}
\description{
# internal function to calculate enrichment in differences in correlation between two groups
# access through the leapr wrapper
}
