% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_comparison_enrichment.R
\name{correlation_comparison_enrichment}
\alias{correlation_comparison_enrichment}
\title{correlation_comparison_enrichment}
\usage{
correlation_comparison_enrichment(geneset, abundance, set1, set2,
  mapping_column = NA, tag = NA, mode = "original")
}
\arguments{
\item{geneset}{is...}

\item{abundance}{is...}

\item{set1}{is...}

\item{set2}{is...}

\item{mapping_column}{default is an NA}

\item{tag}{default is an NA}

\item{mode}{default is 'original'}
}
\description{
correlation_comparison_enrichment function description is...
}
\examples{
dontrun{


}

}
