% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_comparison_enrichment.R
\name{correlation_comparison_enrichment}
\alias{correlation_comparison_enrichment}
\title{correlation_comparison_enrichment}
\usage{
correlation_comparison_enrichment(geneset, abundance, set1, set2,
  mapping_column = NA, tag = NA, mode = "original")
}
\arguments{
\item{geneset}{is a list of four vectors, gene names, gene descriptions, gene sizes and a matrix...??}

\item{abundance}{Is a \emph{mxn} matrix of gene expression data, with \emph{m} gene names (rows) and \emph{n} sample/condition (columns).}

\item{set1}{Is a character vector composed of column names from \code{abundance}, that ...??.}

\item{set2}{Is a character vector of column names, similar to \code{set1} ...??.}

\item{mapping_column}{Is a character string, a column name of \code{abundance}, that...??.}

\item{tag}{default is an NA}

\item{mode}{default is 'original'}
}
\description{
correlation_comparison_enrichment function description is...
}
\examples{
dontrun{


}

}
