% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_enrichment.R
\name{correlation_enrichment}
\alias{correlation_enrichment}
\title{correlation_enrichment}
\usage{
correlation_enrichment(geneset, eset, assay_name, mapping_column = NA)
}
\arguments{
\item{geneset}{Geneset list}

\item{eset}{a SummarizedExperiment object}

\item{assay_name}{name of assay}

\item{mapping_column}{Column to use to map identifiers, if not rownames}
}
\value{
list of enrichment statistic table and correlation matrix
}
\description{
# calculate enrichment in correlation between pathway members
# access through leapr wrapper
}
