% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_enrichment.R
\name{correlation_enrichment}
\alias{correlation_enrichment}
\title{correlation_enrichment}
\usage{
correlation_enrichment(geneset, abundance, mapping_column = NA, tag = NA)
}
\arguments{
\item{geneset}{Geneset list}

\item{abundance}{Abundance value matrix}

\item{mapping_column}{Column to use to map}

\item{tag}{tag to append to group name}
}
\value{
list of enrichment statistic table and correlation matrix
}
\description{
# calculate enrichment in correlation between pathway members
# access through leapr wrapper
}
