% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_enrichment.R
\name{correlation_enrichment}
\alias{correlation_enrichment}
\title{correlation_enrichment}
\usage{
correlation_enrichment(geneset, abundance, mapping_column = NA,
  tag = NA)
}
\arguments{
\item{geneset}{is...}

\item{abundance}{is...}

\item{mapping_column}{default is NA}

\item{tag}{default is NA}
}
\description{
correlation_enrichment function description is...
}
\examples{
dontrun{


}

}
