% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_enrichment.R
\name{correlation_enrichment}
\alias{correlation_enrichment}
\title{correlation_enrichment}
\usage{
correlation_enrichment(geneset, abundance, mapping_column = NA, tag = NA)
}
\description{
# calculate enrichment in correlation between pathway members
# access through leapr wrapper
}
