% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_by_fishers.R
\name{enrichment_by_fishers}
\alias{enrichment_by_fishers}
\title{enrichment_by_fishers}
\usage{
enrichment_by_fishers(group, background, annotation)
}
\arguments{
\item{group}{is a list of identifiers (e.g. differentially expressed gene ids) to test for enrichment}

\item{background}{is the total list of identifiers to consider as a background}

\item{annotation}{is a list of identifiers to compare with group (e.g. a pathway)}
}
\description{
Calculate statistical enrichment by membership (Fisher's exact test) between a group and background
based on an annotation.
}
\details{
This function uses the Fisher's exact test to test a contingency table formed by the group, the annotation,

and the background - a lists of identifiers. Also known as a hypergeometric test, this essentially tests

how surprising it is that one group (group) overlaps with another group (annotation) in the context of the

background of all identifiers considered (background; e.g. all the proteins identified in the assay), to the

extent observed in a dataset.
}
\examples{
dontrun{

 background = c("A","B","C","D","E","F","G","H","I","J","K","L")
 group = c("A","B","C","D")
 annotation = c("A", "B","C","D","E")
 
 result = enrichment_by_fishers(group, background, annotation)
 result$fisher$p.value


}

}
