% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_by_fishers.R
\name{enrichment_by_fishers}
\alias{enrichment_by_fishers}
\title{enrichment_by_fishers}
\usage{
enrichment_by_fishers(group, background, annotation)
}
\description{
# helper function for enrichment functions
}
