% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_by_fishers.R
\name{enrichment_by_fishers}
\alias{enrichment_by_fishers}
\title{enrichment_by_fishers}
\usage{
enrichment_by_fishers(group, background, annotation)
}
\arguments{
\item{group}{is...}

\item{background}{Is a \emph{mxn} matrix of gene expression data, with \emph{m} gene names (rows) and \emph{n} sample/condition (columns).}

\item{annotation}{is...}
}
\description{
enrichment_by_fishers function description is...
}
\examples{
dontrun{


}

}
