% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_comparison.R
\name{enrichment_comparison}
\alias{enrichment_comparison}
\title{enrichment_comparison}
\usage{
enrichment_comparison(
  geneset,
  abundance,
  mapping_column = NULL,
  abundance_column = NULL,
  fdr = 0,
  matchset = NULL,
  longform = F,
  sample_comparison,
  min_p_threshold = NULL,
  tag = NA,
  sample_n = NULL
)
}
\arguments{
\item{geneset}{is a GeneSet object for pathway annotation}

\item{mapping_column}{Is an optional character string, a column name of \code{abundance}, for specifying gene mapping (e.g. for phosphoproteomics data). Defaults to NULL}

\item{abundance_column}{Is a character vector composed of column names from \code{abundance}, that identifies which conditions should be tested.}

\item{fdr}{A numerical value which specifies how many times to randomly sample genes, defaults to 0.}

\item{matchset}{defaults to NULL}

\item{sample_comparison}{Is a character vector composed of column names from \code{abundance}, that identifies which conditions should be compared.}

\item{min_p_threshold}{Is a numeric value, a lower p-value threshold for returning results, defaults to NULL.}

\item{sample_n}{If set specifies the number of randomizations to perform to calculate an FDR [warning: time intensive], default is NA}

\item{Is}{a \emph{mxn} matrix of abundance data (protein, gene, etc.), with \emph{m} gene names (rows) and \emph{n} sample/condition (columns).}

\item{a}{prefix tag (optional) for specifying source of data default is NA}
}
\description{
Enrichment comparison is a wrapper function for enrichment_in_abundance where 'sample_comparison' argument is a required input.
}
