% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_abundance.R
\name{enrichment_in_abundance}
\alias{enrichment_in_abundance}
\title{enrichment_in_abundance}
\usage{
enrichment_in_abundance(
  geneset,
  abundance,
  mapping_column = NULL,
  abundance_column = NULL,
  fdr = 0,
  matchset = NULL,
  longform = F,
  sample_comparison = NULL,
  min_p_threshold = NULL,
  tag = NA,
  sample_n = NULL,
  silence_try_errors = T
)
}
\description{
Enrichment in abundance calculates enrichment in pathways by the difference in abundance of the pathway members.
}
