% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_abundance.R
\name{enrichment_in_abundance}
\alias{enrichment_in_abundance}
\title{enrichment_in_abundance}
\usage{
enrichment_in_abundance(geneset, abundance, mapping_column = NULL,
  abundance_column = NULL, fdr = 0, matchset = NULL, longform = F,
  sample_comparison = NULL, background_comparison = NULL,
  min_p_threshold = NULL, tag = NA, sample_n = NULL)
}
\arguments{
\item{geneset}{is...}

\item{abundance}{is...}

\item{mapping_column}{defaults to NULL}

\item{abundance_column}{defaults to NULL}

\item{fdr}{defaults to 0}

\item{matchset}{defaults to NULL}

\item{longform}{is a logical. Defaults to FALSE}

\item{sample_comparison}{defaults to NULL}

\item{background_comparison}{defaults to NULL}

\item{min_p_threshold}{defaults to NULL}

\item{tag}{defaults to an NA}

\item{sample_n}{defaults to NULL}
}
\description{
enrichment_in_abundance function description is...
}
\examples{
dontrun{


}

}
