% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_groups.R
\name{enrichment_in_groups}
\alias{enrichment_in_groups}
\title{enrichment_in_groups}
\usage{
enrichment_in_groups(geneset, targets = NULL, background = NULL,
  method = "fishers", minsize = 5, mapping_column = NULL,
  abundance_column = NULL, randomize = F)
}
\arguments{
\item{targets}{defaults to NULL}

\item{background}{defaults to NULL}

\item{method}{defaults to 'fishers'}

\item{minsize}{defaults to 5}

\item{mapping_column}{defaults to NULL}

\item{abundance_column}{defaults to NULL}

\item{randomize}{is a logical, defaults to FALSE}

\item{genesets}{is...}
}
\description{
enrichment_in_groups function description is...
}
\examples{
dontrun{


}

}
