% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_order.R
\name{enrichment_in_order}
\alias{enrichment_in_order}
\title{enrichment_in_order}
\usage{
enrichment_in_order(
  geneset,
  background = NULL,
  minsize = 5,
  mapping_column = NULL,
  abundance_column = NULL,
  randomize = F
)
}
\description{
Enrichment in order is a wrapper function for enrichment_in_groups where 'method' argument is set to "ks" and "targets" is set to NA
}
