% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_relationships.R
\name{enrichment_in_relationships}
\alias{enrichment_in_relationships}
\title{enrichment_in_relationships}
\usage{
enrichment_in_relationships(
  geneset,
  relationships,
  idmap = NA,
  tag = NA,
  mode = "original"
)
}
\description{
enrichment_in_relationships function description is a general way to determine if a pathway
is enriched in relationships (interactions, correlation) between its members
# access through leapr wrapper
}
