% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_in_relationships.R
\name{enrichment_in_relationships}
\alias{enrichment_in_relationships}
\title{enrichment_in_relationships}
\usage{
enrichment_in_relationships(geneset, relationships, idmap = NA,
  tag = NA, mode = "original")
}
\arguments{
\item{geneset}{is...}

\item{relationships}{is...}

\item{idmap}{defaults to an NA}

\item{tag}{defaults to an NA}

\item{mode}{defaults to 'original'}
}
\description{
enrichment_in_relationships function description is...
}
\examples{
dontrun{


}

}
