% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathway_information.R
\name{get_pathway_information}
\alias{get_pathway_information}
\title{get_pathway_information}
\usage{
get_pathway_information(geneset, path, remove.tags = FALSE)
}
\arguments{
\item{geneset}{is a GeneSet object for pathway annotation}

\item{path}{is the name of the gene set pathway to be return}

\item{remove.tags}{boolean indicating whether ot remove tags}
}
\value{
list of pathway information
}
\description{
get_pathway_information extracts information about a pathway from a GeneSet object
}
\examples{

     library(leapR)
     
     # load example gene set
     data("ncipid")
     
     tnfpathway = get_pathway_information(ncipid, "tnfpathway")


}
