% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathway_information.R
\name{get_pathway_information}
\alias{get_pathway_information}
\title{get_pathway_information}
\usage{
get_pathway_information(geneset, path, remove.tags = F)
}
\arguments{
\item{geneset}{is a GeneSet object for pathway annotation}

\item{name}{is the name of the gene set to be returned}
}
\description{
get_pathway_information extracts information about a pathway from a GeneSet object
}
\examples{
dontrun{
     library(leapr)
     
     # load example gene set
     data("ncipid")
     
     tnfpathway = get_pathway_information(ncipid, "tnfpathway")

}

}
