% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapR-package.R
\docType{package}
\name{leapR-package}
\alias{leapR-package}
\title{leapR: Layered enrichment analysis of pathways R}
\description{
leapR is a package that identifies pathways that are enriched across diverse 'omics experiments. It leverages any tabular expression data (proteomics, transcriptomics) using the `SummarizedExperiment` object. It works with any pathway in the .gct file format.
}
\author{
\strong{Maintainer}: Sara Gosline \email{sara.gosline@pnnl.gov} (\href{https://orcid.org/0000-0002-6534-4774}{ORCID})

Authors:
\itemize{
  \item Jason McDermott \email{jason.mcdermott@pnnl.gov}
  \item Jeremy Jacobson \email{jeremy.jacobson@pnnl.gov}
}

Other contributors:
\itemize{
  \item Vincent Danna \email{vincent.danna@pnnl.gov} [contributor]
  \item National Institutes of Health [funder]
}

}
\keyword{internal}
