% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msigdb}
\alias{msigdb}
\title{MSIGDB
Is this necessary?}
\format{
An object of class \code{geneset_data} (inherits from \code{list}) of length 4.
}
\usage{
msigdb
}
\description{
MSIGDB
Is this necessary?
}
\keyword{datasets}
