% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncipid}
\alias{ncipid}
\title{NCI Gene lists}
\format{
A list with 4 items
\describe{
\item{names} The names of the signling pathways
\item{desc}
\item{sizes}
\item{matrix}
}
}
\source{
Where did this come from?
}
\usage{
ncipid
}
\description{
A list of pathways and the genes that comprise these pathways
}
\keyword{datasets}
