% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_sets.R
\name{read_gene_sets}
\alias{read_gene_sets}
\title{read_gene_sets}
\usage{
read_gene_sets(gsfile, gene.labels = NA, gs.size.threshold.min = 5,
  gs.size.threshold.max = 15000)
}
\arguments{
\item{gsfile}{is...}

\item{gene.labels}{defaults to NA}

\item{gs.size.threshold.min}{defaults to 5}

\item{gs.size.threshold.max}{defaults to 15000}
}
\description{
read_gene_sets function description is...
}
\examples{
dontrun{


}

}
