% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gene_sets.R
\name{read_gene_sets}
\alias{read_gene_sets}
\title{read_gene_sets}
\usage{
read_gene_sets(
  gsfile,
  gene.labels = NA,
  gs.size.threshold.min = 5,
  gs.size.threshold.max = 15000
)
}
\arguments{
\item{gsfile}{is a gene set file, for example a .gmt file (gene matrix transposed file format)}

\item{gene.labels}{defaults to NA}

\item{gs.size.threshold.min}{defaults to 5}

\item{gs.size.threshold.max}{defaults to 15000}
}
\description{
read_gene_sets function description is...
}
\examples{
dontrun{
        library(readr)
        data("protdata")

        #we have to update the colnames slightly to match what we have in the patient groups
        colnames(protdata) = sapply(colnames(protdata), function (r) {a=strsplit(r,"\\\\.");paste("TCGA",a[[1]][2], a[[1]][3], sep="-")})

        #read in the pathways
        ncipid = read_gene_sets(gsfile = "/example/NCI_PID_genesymbol_corrected.gmt")

}

}
