% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shortlist}
\alias{shortlist}
\title{A list of pathways and genes that comprise these pathways from msigdb}
\format{
a list with 4 items

Short list of patient samples
}
\usage{
shortlist
}
\description{
A list of pathways and genes that comprise these pathways from msigdb
}
\keyword{datasets}
