% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transdata}
\alias{transdata}
\title{Transcriptomic data}
\format{
A data frame with 18632 rows and 174 columns
}
\source{
Where did this come from?
}
\usage{
transdata
}
\description{
A dataset containing gene transcriptomic data containing 18632 genes and 174 patient samples. 
Values represent normalized (Z score) gene expression relative to the mean expression for that gene
}
\keyword{datasets}
